+++ src/main/java/wooteco/subway/admin/controller/LineController.java
@@ -1,21 +1,17 @@
 package wooteco.subway.admin.controller;
 
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import wooteco.subway.admin.domain.Line;
 import wooteco.subway.admin.domain.Station;
-import wooteco.subway.admin.dto.ErrorResponse;
 import wooteco.subway.admin.dto.LineRequest;
 import wooteco.subway.admin.dto.LineResponse;
 import wooteco.subway.admin.dto.LineStationCreateRequest;
-import wooteco.subway.admin.exception.StartStationNotFoundException;
 import wooteco.subway.admin.service.LineService;
 import wooteco.subway.admin.service.StationService;
 
 import javax.validation.Valid;
 import java.net.URI;
-import java.sql.SQLException;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Set;
@@ -79,24 +75,6 @@ public class LineController {
     @DeleteMapping("/lines/{lineId}/stations/{stationId}")
     public void deleteLineStation(@PathVariable Long lineId, @PathVariable Long stationId) {
         lineService.removeLineStation(lineId, stationId);
-    }
-
-    @ExceptionHandler(SQLException.class)
-    public ResponseEntity<ErrorResponse> handleSQLException(SQLException e) {
-        final ErrorResponse response = ErrorResponse.of("SQL 에러입니다.");
-        return new ResponseEntity<>(response, HttpStatus.valueOf(500));
-    }
-
-    @ExceptionHandler(StartStationNotFoundException.class)
-    public ResponseEntity<ErrorResponse> handleStartStationException(StartStationNotFoundException e) {
-        final ErrorResponse response = ErrorResponse.of("시작역을 찾을 수 없습니다.");
-        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
-    }
-
-    @ExceptionHandler(IllegalArgumentException.class)
-    public ResponseEntity<ErrorResponse> handleIllegalArgumentException(StartStationNotFoundException e) {
-        final ErrorResponse response = ErrorResponse.of(e.getMessage());
-        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
     }
 
     private LineResponse findLineWithStationsById(Long id) {
